// src/app/shared/generated/model/applicationParameterHistory.ts
/**
 * onecx-parameters-bff
 * OneCx parameters Bff
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApplicationParameterHistory { 
    id?: string;
    modificationCount?: number;
    creationDate?: string;
    creationUser?: string;
    modificationDate?: string;
    modificationUser?: string;
    applicationId?: string;
    productName?: string;
    key?: string;
    usedValue?: string;
    defaultValue?: string;
    type?: string;
    instanceId?: string;
}